scripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.114
botocore==1.29.114
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.27.4
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
